Station

+++ EF-Exams/StationAuthor/Stations.App/Startup.cs
@@ -1,6 +1,8 @@
 ﻿using System;
+using System.Data.SqlClient;
 using System.IO;
 using AutoMapper;
+using Microsoft.EntityFrameworkCore;
 using Stations.Data;
 
 namespace Stations.App
@@ -17,13 +19,12 @@ namespace Stations.App
 			Mapper.Initialize(cfg => cfg.AddProfile<StationsProfile>());
 
 			ImportEntities(context);
-		    
 			ExportEntities(context);
 		}
 
-		private static void ImportEntities(StationsDbContext context, string baseDir = @"Datasets\")
+		private static void ImportEntities(StationsDbContext context, string baseDir = @"..\Datasets\")
 		{
-			const string exportDir = "ImportResults/";
+			const string exportDir = "./ImportResults/";
 
 			var stations = DataProcessor.Deserializer.ImportStations(context, File.ReadAllText(baseDir + "stations.json"));
 			PrintAndExportEntityToFile(stations, exportDir + "Stations.txt");
@@ -46,13 +47,13 @@ namespace Stations.App
 
 		private static void ExportEntities(StationsDbContext context)
 		{
-			const string exportDir = "ImportResults/";
+			const string exportDir = "./ImportResults/";
 
 			var jsonOutput = DataProcessor.Serializer.ExportDelayedTrains(context, "01/01/2017");
 			Console.WriteLine(jsonOutput);
 			File.WriteAllText(exportDir + "DelayedTrains.json", jsonOutput);
 
-			var xmlOutput = DataProcessor.Serializer.ExportCardsTicket(context, "Elder");
+			var xmlOutput = DataProcessor.Serializer.ExportCardsTicket(context, "Debilitated");
 			Console.WriteLine(xmlOutput);
 			File.WriteAllText(exportDir + "CardsTicket.xml", xmlOutput);
 		}
@@ -63,10 +64,33 @@ namespace Stations.App
 			File.WriteAllText(outputPath, entityOutput.TrimEnd());
 		}
 
-		private static void ResetDatabase(StationsDbContext context)
+		private static void ResetDatabase(StationsDbContext context, bool shouldDeleteDatabase = false)
 		{
-			context.Database.EnsureDeleted();
+			if (shouldDeleteDatabase)
+			{
+				context.Database.EnsureDeleted();
+				context.Database.EnsureCreated();
+			}
+
 			context.Database.EnsureCreated();
+
+			var disableIntegrityChecksQuery = "EXEC sp_MSforeachtable @command1='ALTER TABLE ? NOCHECK CONSTRAINT ALL'";
+			context.Database.ExecuteSqlCommand(disableIntegrityChecksQuery);
+
+			var deleteRowsQuery = "EXEC sp_MSforeachtable @command1='DELETE FROM ?'";
+			context.Database.ExecuteSqlCommand(deleteRowsQuery);
+
+			var enableIntegrityChecksQuery = "EXEC sp_MSforeachtable @command1='ALTER TABLE ? WITH CHECK CHECK CONSTRAINT ALL'";
+			context.Database.ExecuteSqlCommand(enableIntegrityChecksQuery);
+
+			var reseedQuery = "EXEC sp_MSforeachtable @command1='DBCC CHECKIDENT(''?'', RESEED, 0)'";
+			try
+			{
+				context.Database.ExecuteSqlCommand(reseedQuery);
+			}
+			catch (SqlException) // OrderItems table has no identity column, which isn't a problem
+			{
+			}
 		}
 	}
 }
